+++ example/chinese_lunar_calendar_example.dart
@@ -11,6 +11,8 @@ void main() {
   print(
       '阴历数字: ${lunarCalendar.lunarDate.year}, ${lunarCalendar.lunarDate.month}, ${lunarCalendar.lunarDate.day}, 闰月：${lunarCalendar.lunarDate.isLeapMonth}');
   print('阴历汉字: ${lunarCalendar.lunarDate.fullCNString}');
+  print(
+      '八字: ${lunarCalendar.lunarDate.lunarYear8Char} ${lunarCalendar.lunarMonth8Char}');
   print('星期：${lunarCalendar.weekDayCN}');
   print('月相：${lunarCalendar.phaseOfMoon}');
   print(