+++ test/utils/calculations_test.dart
@@ -74,6 +74,13 @@ void main() {
       expect(getLunarYear8Char(lunarYear: 1670), '庚戌');
     });
 
+    test('Test getLunarMonth8Char', () {
+      expect(
+          getLunarMonth8Char(
+              dateTime: DateTime(2023, 6, 24), useCSTToCalculate: false),
+          '戊午');
+    });
+
     test('Test getWeekDayCN', () {
       expect(getWeekDayCN(DateTime(2023, 6, 4)), '星期日');