+++ CHANGELOG.md
@@ -17,3 +17,4 @@
 - Add option for calculating lunar calendar using China Standard Time or local time
 
 ## 0.0.6
+- Calculate month 8 char.